feat: public key export
