+++ src/composed/key/secret.rs
@@ -1,129 +1,72 @@
-use std::io;
+use chrono;
 
-use armor;
-use composed::key::{SignedKeyDetails, SignedPublicSubKey};
-use crypto::hash::HashAlgorithm;
-use crypto::public_key::PublicKeyAlgorithm;
+use composed::{KeyDetails, PublicSubkey, SignedSecretKey, SignedSecretSubKey};
+use crypto::PublicKeyAlgorithm;
 use errors::Result;
-use packet::{self, write_packet, SignatureType};
-use ser::Serialize;
-use types::{KeyId, KeyTrait, PublicKeyTrait, SecretKeyRepr, SecretKeyTrait};
+use packet::{self, KeyFlags, SignatureConfigBuilder, SignatureType, Subpacket};
+use types::{KeyId, KeyTrait, SecretKeyTrait};
 
-/// Represents a secret signed PGP key.
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct SignedSecretKey {
-    pub primary_key: packet::SecretKey,
-    pub details: SignedKeyDetails,
-    pub public_subkeys: Vec<SignedPublicSubKey>,
-    pub secret_subkeys: Vec<SignedSecretSubKey>,
+/// User facing interface to work with a secret key.
+#[derive(Debug, PartialEq, Eq)]
+pub struct SecretKey {
+    primary_key: packet::SecretKey,
+    details: KeyDetails,
+    public_subkeys: Vec<PublicSubkey>,
+    secret_subkeys: Vec<SecretSubkey>,
 }
 
-key_parser!(
-    SignedSecretKey,
-    SignedSecretKeyParser,
-    Tag::SecretKey,
-    packet::SecretKey,
-    // secret keys, can contain both public and secret subkeys
-    (
-        PublicSubkey,
-        packet::PublicSubkey,
-        SignedPublicSubKey,
-        public_subkeys
-    ),
-    (
-        SecretSubkey,
-        packet::SecretSubkey,
-        SignedSecretSubKey,
-        secret_subkeys
-    )
-);
+#[derive(Debug, PartialEq, Eq)]
+pub struct SecretSubkey {
+    key: packet::SecretSubkey,
+    keyflags: KeyFlags,
+}
 
-impl SignedSecretKey {
+impl SecretKey {
     pub fn new(
         primary_key: packet::SecretKey,
-        details: SignedKeyDetails,
-        public_subkeys: Vec<SignedPublicSubKey>,
-        secret_subkeys: Vec<SignedSecretSubKey>,
+        details: KeyDetails,
+        public_subkeys: Vec<PublicSubkey>,
+        secret_subkeys: Vec<SecretSubkey>,
     ) -> Self {
-        let public_subkeys = public_subkeys
-            .into_iter()
-            .filter(|key| {
-                if key.signatures.is_empty() {
-                    warn!("ignoring unsigned {:?}", key.key);
-                    false
-                } else {
-                    true
-                }
-            })
-            .collect();
-
-        let secret_subkeys = secret_subkeys
-            .into_iter()
-            .filter(|key| {
-                if key.signatures.is_empty() {
-                    warn!("ignoring unsigned {:?}", key.key);
-                    false
-                } else {
-                    true
-                }
-            })
-            .collect();
-
-        SignedSecretKey {
+        SecretKey {
             primary_key,
             details,
             public_subkeys,
             secret_subkeys,
         }
     }
-    fn verify_public_subkeys(&self) -> Result<()> {
-        for subkey in &self.public_subkeys {
-            subkey.verify(&self.primary_key)?;
-        }
 
-        Ok(())
-    }
+    pub fn sign<F>(self, key_pw: F) -> Result<SignedSecretKey>
+    where
+        F: (FnOnce() -> String) + Clone,
+    {
+        let primary_key = self.primary_key;
+        let details = self.details.sign(&primary_key, key_pw.clone())?;
+        let public_subkeys = self
+            .public_subkeys
+            .into_iter()
+            .map(|k| k.sign(&primary_key, key_pw.clone()))
+            .collect::<Result<Vec<_>>>()?;
+        let secret_subkeys = self
+            .secret_subkeys
+            .into_iter()
+            .map(|k| k.sign(&primary_key, key_pw.clone()))
+            .collect::<Result<Vec<_>>>()?;
 
-    fn verify_secret_subkeys(&self) -> Result<()> {
-        for subkey in &self.secret_subkeys {
-            subkey.verify(&self.primary_key)?;
-        }
-
-        Ok(())
-    }
-
-    pub fn verify(&self) -> Result<()> {
-        self.details.verify(&self.primary_key)?;
-        self.verify_public_subkeys()?;
-        self.verify_secret_subkeys()?;
-
-        Ok(())
-    }
-
-    pub fn to_armored_writer(&self, writer: &mut impl io::Write) -> Result<()> {
-        armor::write(self, "PRIVATE KEY", writer)
-    }
-
-    pub fn to_armored_bytes(&self) -> Result<Vec<u8>> {
-        let mut buf = Vec::new();
-
-        self.to_armored_writer(&mut buf)?;
-
-        Ok(buf)
-    }
-
-    pub fn to_armored_string(&self) -> Result<String> {
-        Ok(::std::str::from_utf8(&self.to_armored_bytes()?)?.to_string())
+        Ok(SignedSecretKey {
+            primary_key,
+            details,
+            public_subkeys,
+            secret_subkeys,
+        })
     }
 }
 
-impl KeyTrait for SignedSecretKey {
-    /// Returns the fingerprint of the associated primary key.
+impl KeyTrait for SecretKey {
     fn fingerprint(&self) -> Vec<u8> {
         self.primary_key.fingerprint()
     }
 
-    /// Returns the Key ID of the associated primary key.
     fn key_id(&self) -> Option<KeyId> {
         self.primary_key.key_id()
     }
@@ -133,149 +76,46 @@ impl KeyTrait for SignedSecretKey {
     }
 }
 
-impl Serialize for SignedSecretKey {
-    fn to_writer<W: io::Write>(&self, writer: &mut W) -> Result<()> {
-        write_packet(writer, &self.primary_key)?;
-        self.details.to_writer(writer)?;
-        for ps in &self.public_subkeys {
-            ps.to_writer(writer)?;
-        }
-
-        for ps in &self.secret_subkeys {
-            ps.to_writer(writer)?;
-        }
-
-        Ok(())
+impl SecretSubkey {
+    pub fn new(key: packet::SecretSubkey, keyflags: KeyFlags) -> Self {
+        SecretSubkey { key, keyflags }
     }
-}
 
-impl SecretKeyTrait for SignedSecretKey {
-    fn unlock<F, G>(&self, pw: F, work: G) -> Result<()>
+    pub fn sign<F>(self, sec_key: &impl SecretKeyTrait, key_pw: F) -> Result<SignedSecretSubKey>
     where
-        F: FnOnce() -> String,
-        G: FnOnce(&SecretKeyRepr) -> Result<()>,
+        F: (FnOnce() -> String) + Clone,
     {
-        self.primary_key.unlock(pw, work)
-    }
+        let key = self.key;
+        let hashed_subpackets = vec![
+            Subpacket::SignatureCreationTime(chrono::Utc::now()),
+            Subpacket::KeyFlags(self.keyflags.into()),
+        ];
 
-    fn create_signature<F>(
-        &self,
-        key_pw: F,
-        hash: HashAlgorithm,
-        data: &[u8],
-    ) -> Result<Vec<Vec<u8>>>
-    where
-        F: FnOnce() -> String,
-    {
-        self.primary_key.create_signature(key_pw, hash, data)
-    }
-}
+        let config = SignatureConfigBuilder::default()
+            .typ(SignatureType::SubkeyBinding)
+            .pub_alg(sec_key.algorithm())
+            .hashed_subpackets(hashed_subpackets)
+            .unhashed_subpackets(vec![
+                Subpacket::Issuer(sec_key.key_id().expect("missing key id")),
+                Subpacket::IssuerFingerprint(sec_key.fingerprint()),
+            ])
+            .build()?;
+        let signatures = vec![config.sign_key_binding(sec_key, key_pw, &key)?];
 
-impl PublicKeyTrait for SignedSecretKey {
-    fn verify_signature(&self, hash: HashAlgorithm, data: &[u8], sig: &[Vec<u8>]) -> Result<()> {
-        self.primary_key.verify_signature(hash, data, sig)
-    }
-
-    fn to_writer_old(&self, _writer: &mut impl io::Write) -> Result<()> {
-        unimplemented!("")
+        Ok(SignedSecretSubKey { key, signatures })
     }
 }
 
-/// Represents a composed secret PGP SubKey.
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct SignedSecretSubKey {
-    pub key: packet::SecretSubkey,
-    pub signatures: Vec<packet::Signature>,
-}
-
-impl SignedSecretSubKey {
-    pub fn new(key: packet::SecretSubkey, signatures: Vec<packet::Signature>) -> Self {
-        let signatures = signatures
-            .into_iter()
-            .filter(|sig| {
-                if sig.typ() != SignatureType::SubkeyBinding
-                    && sig.typ() != SignatureType::SubkeyRevocation
-                {
-                    warn!(
-                        "ignoring unexpected signature {:?} after Subkey packet",
-                        sig.typ()
-                    );
-                    false
-                } else {
-                    true
-                }
-            })
-            .collect();
-
-        SignedSecretSubKey { key, signatures }
-    }
-
-    pub fn verify(&self, key: &impl PublicKeyTrait) -> Result<()> {
-        ensure!(!self.signatures.is_empty(), "missing subkey bindings");
-
-        for sig in &self.signatures {
-            sig.verify_key_binding(key, &self.key)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl KeyTrait for SignedSecretSubKey {
-    /// Returns the fingerprint of the key.
+impl KeyTrait for SecretSubkey {
     fn fingerprint(&self) -> Vec<u8> {
         self.key.fingerprint()
     }
 
-    /// Returns the Key ID of the key.
     fn key_id(&self) -> Option<KeyId> {
         self.key.key_id()
     }
 
     fn algorithm(&self) -> PublicKeyAlgorithm {
         self.key.algorithm()
-    }
-}
-
-impl Serialize for SignedSecretSubKey {
-    fn to_writer<W: io::Write>(&self, writer: &mut W) -> Result<()> {
-        write_packet(writer, &self.key)?;
-        for sig in &self.signatures {
-            write_packet(writer, sig)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl SecretKeyTrait for SignedSecretSubKey {
-    fn unlock<F, G>(&self, pw: F, work: G) -> Result<()>
-    where
-        F: FnOnce() -> String,
-        G: FnOnce(&SecretKeyRepr) -> Result<()>,
-    {
-        self.key.unlock(pw, work)
-    }
-
-    fn create_signature<F>(
-        &self,
-        key_pw: F,
-        hash: HashAlgorithm,
-        data: &[u8],
-    ) -> Result<Vec<Vec<u8>>>
-    where
-        F: FnOnce() -> String,
-    {
-        self.key.create_signature(key_pw, hash, data)
-    }
-}
-
-impl PublicKeyTrait for SignedSecretSubKey {
-    fn verify_signature(&self, hash: HashAlgorithm, data: &[u8], sig: &[Vec<u8>]) -> Result<()> {
-        self.key.verify_signature(hash, data, sig)
-    }
-
-    fn to_writer_old(&self, _writer: &mut impl io::Write) -> Result<()> {
-        unimplemented!("")
     }
 }
