+++ src/composed/mod.rs
@@ -1,11 +1,10 @@
 pub mod key;
 pub mod message;
+pub mod signed_key;
 
 mod message_parser;
 mod shared;
 
-pub use self::key::{
-    SignedKeyDetails, SignedPublicKey, SignedPublicSubKey, SignedSecretKey, SignedSecretSubKey,
-};
-pub use self::message::Message;
+pub use self::key::*;
 pub use self::shared::Deserializable;
+pub use self::signed_key::*;
