+++ src/composed/signed_key/mod.rs
@@ -1,0 +1,12 @@
+#[macro_use]
+mod key_parser_macros;
+
+pub mod parse;
+pub mod public;
+pub mod secret;
+pub mod shared;
+
+pub use self::parse::*;
+pub use self::public::*;
+pub use self::secret::*;
+pub use self::shared::*;
