+++ src/crypto/mod.rs
@@ -3,6 +3,8 @@ pub mod aes_kw;
 pub mod checksum;
 pub mod ecc;
 pub mod ecc_curve;
+pub mod ecdh;
+pub mod eddsa;
 pub mod hash;
 pub mod kdf;
 pub mod pkcs5;
@@ -10,3 +12,18 @@ pub mod public_key;
 pub mod rsa;
 pub mod signature;
 pub mod sym;
+
+pub use self::aead::*;
+pub use self::aes_kw::*;
+pub use self::checksum::*;
+pub use self::ecc::*;
+pub use self::ecc_curve::*;
+pub use self::ecdh::*;
+pub use self::eddsa::*;
+pub use self::hash::*;
+pub use self::kdf::*;
+pub use self::pkcs5::*;
+pub use self::public_key::*;
+pub use self::rsa::*;
+pub use self::signature::*;
+pub use self::sym::*;
