+++ src/types/key.rs
@@ -1,8 +1,7 @@
 use crypto::public_key::PublicKeyAlgorithm;
-use ser::Serialize;
 use types::KeyId;
 
-pub trait KeyTrait: Serialize + ::std::fmt::Debug {
+pub trait KeyTrait: ::std::fmt::Debug {
     fn fingerprint(&self) -> Vec<u8>;
 
     /// Returns the Key ID of the associated primary key.
